Name truncation: exclude whitespace/punctuation
Previously, if you had a meeting called "ABC DEF GHI" and truncated it
to four characters long, it would appear as "ABC ...". Now it shows as
"ABC...", which is the expected behavior.

This commit also fixes a bug that caused names to be truncated to one
character longer than the length slider's actual position.