Update Workplace video call url regex
This change makes the regex for workplace calls more specific. This prevents an issue where other workplace links (e.g. to posts, or notes) are picked up as the video call link.

I had tested this before using the custom regex setting in preferences. And that works fine, but does not show the workplace icon. Which is why I want to upstream the improved logic